Refactor project into folders
* Also moved the module base to the base of the repository to allow
direct installation of the module from Blender interface rather than
it having to be manual by copying the gta folder (or zipping the gta
folder). This also enables the project installation folder name to be
more unique than "gta".
