+++ .gitignore
@@ -1,2 +1,2 @@
-gta/__pycache__
-gta/GPUCache
+__pycache__
+GPUCache
