+++ __init__.py
@@ -15,7 +15,8 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import bpy
-from . import gui, map_importer
+from .gui import gui
+from .ops import map_importer
 
 from bpy.utils import register_class, unregister_class
