--- gta/gui.py
@@ -1,910 +1,0 @@
-# GTA DragonFF - Blender scripts to edit basic GTA formats
-# Copyright (C) 2019  Parik
-
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-import bpy
-import os
-from . import dff_importer, dff_exporter, col_importer, col_exporter
-from .importer_common import game_version
-from . import map_data
-
-from bpy_extras.io_utils import ImportHelper, ExportHelper
-
-#######################################################
-class EXPORT_OT_dff(bpy.types.Operator, ExportHelper):
-    
-    bl_idname      = "export_dff.scene"
-    bl_description = "Export a Renderware DFF or COL File"
-    bl_label       = "DragonFF DFF (.dff)"
-    filename_ext   = ".dff"
-
-    filepath       : bpy.props.StringProperty(name="File path",
-                                              maxlen=1024,
-                                              default="",
-                                              subtype='FILE_PATH')
-    
-    filter_glob    : bpy.props.StringProperty(default="*.dff;*.col",
-                                              options={'HIDDEN'})
-    
-    directory      : bpy.props.StringProperty(maxlen=1024,
-                                              default="",
-                                              subtype='FILE_PATH')
-
-    mass_export     :  bpy.props.BoolProperty(
-        name        = "Mass Export",
-        default     = False
-    )
-
-    export_coll     : bpy.props.BoolProperty(
-        name        = "Export Collision",
-        default     = True
-    )
-    
-    only_selected   :  bpy.props.BoolProperty(
-        name        = "Only Selected",
-        default     = False
-    )
-    reset_positions :  bpy.props.BoolProperty(
-        name        = "Preserve Positions",
-        description = "Don't set object positions to (0,0,0)",
-        default     = False
-    )
-    export_version  : bpy.props.EnumProperty(
-        items =
-        (
-            ('0x33002', "GTA 3 (v3.3.0.2)", "Grand Theft Auto 3 PC (v3.3.0.2)"),
-            ('0x34003', "GTA VC (v3.4.0.3)", "Grand Theft Auto VC PC (v3.4.0.3)"),
-            ('0x36003', "GTA SA (v3.6.0.3)", "Grand Theft Auto SA PC (v3.6.0.3)"),
-            ('custom', "Custom", "Custom RW Version")
-        ),
-        name = "Version Export"
-    )
-    custom_version      : bpy.props.StringProperty(
-        maxlen=7,
-        default="",
-        name = "Custom Version")
-
-    #######################################################
-    def verify_rw_version(self):
-        if len(self.custom_version) != 7:
-            return False
-
-        for i, char in enumerate(self.custom_version):
-            if i % 2 == 0 and not char.isdigit():
-                return False
-            if i % 2 == 1 and not char == '.':
-                return False
-
-        return True
-    
-    #######################################################
-    def draw(self, context):
-        layout = self.layout
-
-        layout.prop(self, "mass_export")
-
-        if self.mass_export:
-            box = layout.box()
-            row = box.row()
-            row.label(text="Mass Export:")
-
-            row = box.row()
-            row.prop(self, "reset_positions")
-
-        layout.prop(self, "only_selected")
-        layout.prop(self, "export_coll")
-        layout.prop(self, "export_version")
-
-        if self.export_version == 'custom':
-            col = layout.column()
-            col.alert = not self.verify_rw_version()
-            icon = "ERROR" if col.alert else "NONE"
-            
-            col.prop(self, "custom_version", icon=icon)
-        return None
-
-    #######################################################
-    def get_selected_rw_version(self):
-
-        if self.export_version != "custom":
-            return int(self.export_version, 0)
-        
-        else:
-            return int(
-                "0x%c%c%c0%c" % (self.custom_version[0],
-                                 self.custom_version[2],
-                                 self.custom_version[4],
-                                 self.custom_version[6]),
-                0)
-    
-    #######################################################
-    def execute(self, context):
-
-        if self.export_version == "custom":
-            if not self.verify_rw_version():
-                self.report({"ERROR_INVALID_INPUT"}, "Invalid RW Version")
-                return {'FINISHED'}
-        
-        dff_exporter.export_dff(
-            {
-                "file_name"      : self.filepath,
-                "directory"      : self.directory,
-                "selected"       : self.only_selected,
-                "mass_export"    : self.mass_export,
-                "version"        : self.get_selected_rw_version(),
-                "export_coll"    : self.export_coll
-            }
-        )
-
-        # Save settings of the export in scene custom properties for later
-        context.scene['dragonff_imported_version'] = self.export_version
-        context.scene['dragonff_custom_version']   = self.custom_version
-            
-        return {'FINISHED'}
-
-    #######################################################
-    def invoke(self, context, event):
-        if 'dragonff_imported_version' in context.scene:
-            self.export_version = context.scene['dragonff_imported_version']
-        if 'dragonff_custom_version' in context.scene:
-            self.custom_version = context.scene['dragonff_custom_version']
-        
-        context.window_manager.fileselect_add(self)
-        return {'RUNNING_MODAL'}
-
-#######################################################
-class IMPORT_OT_dff(bpy.types.Operator, ImportHelper):
-    
-    bl_idname      = "import_scene.dff"
-    bl_description = 'Import a Renderware DFF or COL File'
-    bl_label       = "DragonFF DFF (.dff)"
-
-    filter_glob   : bpy.props.StringProperty(default="*.dff;*.col",
-                                              options={'HIDDEN'})
-    
-    directory     : bpy.props.StringProperty(maxlen=1024,
-                                              default="",
-                                              subtype='FILE_PATH',
-                                              options={'HIDDEN'})
-    
-    # Stores all the file names to read (not just the firsst)
-    files : bpy.props.CollectionProperty(
-        type    = bpy.types.OperatorFileListElement,
-        options = {'HIDDEN'}
-    )
-
-    # Stores a single file path
-    filepath : bpy.props.StringProperty(
-         name        = "File Path",
-         description = "Filepath used for importing the DFF/COL file",
-         maxlen      = 1024,
-         default     = "",
-         options     = {'HIDDEN'}
-     )
-
-    
-    load_txd :  bpy.props.BoolProperty(
-        name        = "Load TXD file",
-        default     = True
-    )
-
-    connect_bones :  bpy.props.BoolProperty(
-        name        = "Connect Bones",
-        description = "Whether to connect bones (not recommended for anim editing)",
-        default     = False
-    )
-
-    read_mat_split  :  bpy.props.BoolProperty(
-        name        = "Read Material Split",
-        description = "Whether to read material split for loading triangles",
-        default     = False
-    )
-
-    load_images : bpy.props.BoolProperty(
-        name    = "Scan for Images",
-        default = True
-    )
-
-    remove_doubles  :  bpy.props.BoolProperty(
-        name        = "Use Edge Split",
-        default     = True
-    )
-    group_materials :  bpy.props.BoolProperty(
-        name        = "Group Similar Materials",
-        default     = True
-    )
-    
-    image_ext : bpy.props.EnumProperty(
-        items =
-        (
-            ("PNG", ".PNG", "Load a PNG image"),
-            ("JPG", ".JPG", "Load a JPG image"),
-            ("JPEG", ".JPEG", "Load a JPEG image"),
-            ("TGA", ".TGA", "Load a TGA image"),
-            ("BMP", ".BMP", "Load a BMP image"),
-            ("TIF", ".TIF", "Load a TIF image"),
-            ("TIFF", ".TIFF", "Load a TIFF image")
-        ),
-        name        = "Extension",
-        description = "Image extension to search textures in"
-    )
-
-    #######################################################
-    def draw(self, context):
-        layout = self.layout
-
-        layout.prop(self, "load_txd")
-        layout.prop(self, "connect_bones")
-        
-        box = layout.box()
-        box.prop(self, "load_images")
-        if self.load_images:
-            box.prop(self, "image_ext")
-        
-        layout.prop(self, "read_mat_split")
-        layout.prop(self, "remove_doubles")
-        layout.prop(self, "group_materials")
-        
-    #######################################################
-    def execute(self, context):
-        
-        for file in [os.path.join(self.directory,file.name) for file in self.files]:
-            if file.endswith(".col"):
-                col_importer.import_col_file(file, os.path.basename(file))
-                            
-            else:
-                # Set image_ext to none if scan images is disabled
-                image_ext = self.image_ext
-                if not self.load_images:
-                    image_ext = None
-                    
-                importer = dff_importer.import_dff(
-                    {
-                        'file_name'      : file,
-                        'image_ext'      : image_ext,
-                        'connect_bones'  : self.connect_bones,
-                        'use_mat_split'  : self.read_mat_split,
-                        'remove_doubles' : self.remove_doubles,
-                        'group_materials': self.group_materials
-                    }
-                )
-
-                if importer.warning != "":
-                    self.report({'WARNING'}, importer.warning)
-
-                version = importer.version
-
-                # Set imported version to scene settings for use later in export.
-                if version in ['0x33002', '0x34003', '0x36003']:
-                    context.scene['dragonff_imported_version'] = version
-                else:
-                    context.scene['dragonff_imported_version'] = "custom"
-                    context.scene['dragonff_custom_version'] = "{}.{}.{}.{}".format(
-                        *(version[i] for i in [2,3,4,6])
-                    ) #convert hex to x.x.x.x format
-                
-        return {'FINISHED'}
-
-    #######################################################
-    def invoke(self, context, event):
-        
-        context.window_manager.fileselect_add(self)
-        return {'RUNNING_MODAL'}
-
-#######################################################
-class EXPORT_OT_col(bpy.types.Operator, ExportHelper):
-    
-    bl_idname      = "export_col.scene"
-    bl_description = "Export a GTA III/VC/SA Collision File"
-    bl_label       = "DragonFF Collision (.col)"
-    filename_ext   = ".col"
-
-    filepath       : bpy.props.StringProperty(name="File path",
-                                              maxlen=1024,
-                                              default="",
-                                              subtype='FILE_PATH')
-    
-    filter_glob    : bpy.props.StringProperty(default="*.col",
-                                              options={'HIDDEN'})
-    
-    directory      : bpy.props.StringProperty(maxlen=1024,
-                                              default="",
-                                              subtype='FILE_PATH')
-
-    only_selected   :  bpy.props.BoolProperty(
-        name        = "Only Selected",
-        default     = False
-    )
-    
-    export_version  : bpy.props.EnumProperty(
-        items =
-        (
-            ('1', "GTA 3/VC (COLL)", "Grand Theft Auto 3 and Vice City (PC) - Version 1"),
-            ('3', "GTA SA PC/Xbox (COL3)", "Grand Theft Auto SA (PC/Xbox) - Version 3"),
-            ('2', "GTA SA PS2 (COL2)", "Grand Theft Auto SA (PS2) - Version 2")
-        ),
-        name = "Version Export"
-    )
-
-    #######################################################
-    def draw(self, context):
-        layout = self.layout
-        layout.prop(self, "export_version")
-        layout.prop(self, "only_selected")
-        return None
-
-    #######################################################
-    def execute(self, context):
-        
-        col_exporter.export_col(
-            {
-                "file_name"      : self.filepath,
-                "version"        : int(self.export_version),
-                "collection"     : None,
-                "memory"         : False,
-                "mass_export"    : True,
-                "only_selected"  : self.only_selected
-            }
-        )
-
-        # Save settings of the export in scene custom properties for later
-        context.scene['dragonff_imported_version_col'] = self.export_version
-            
-        return {'FINISHED'}
-
-    #######################################################
-    def invoke(self, context, event):
-        if 'dragonff_imported_version_col' in context.scene:
-            self.export_version = context.scene['dragonff_imported_version_col']
-        
-        context.window_manager.fileselect_add(self)
-        return {'RUNNING_MODAL'}
-    
-#######################################################
-class MATERIAL_PT_dffMaterials(bpy.types.Panel):
-
-    bl_idname      = "MATERIAL_PT_dffMaterials"
-    bl_label       = "DragonFF - Export Material"
-    bl_space_type  = "PROPERTIES"
-    bl_region_type = "WINDOW"
-    bl_context     = "material"
-
-    ambient     :  bpy.props.BoolProperty(
-        name        = "Export Material",
-        default     = False
-    )
-
-    #######################################################
-    def draw_col_menu(self, context):
-
-        layout = self.layout
-        settings = context.material.dff
-
-        props = [["col_mat_index", "Material"],
-                 ["col_flags", "Flags"],
-                 ["col_brightness", "Brightness"],
-                 ["col_light", "Light"]]
-        
-        for prop in props:
-            self.draw_labelled_prop(layout.row(), settings, [prop[0]], prop[1])
-
-    #######################################################
-    def draw_labelled_prop(self, row, settings, props, label, text=""):
-
-        row.label(text=label)
-        for prop in props:
-            row.prop(settings, prop, text=text)
-        
-    #######################################################
-    def draw_env_map_box(self, context, box):
-
-        settings = context.material.dff
-
-        box.row().prop(context.material.dff, "export_env_map")
-        if settings.export_env_map:
-            box.row().prop(settings, "env_map_tex", text="Texture")
-
-            self.draw_labelled_prop(
-                box.row(), settings, ["env_map_coef"], "Coefficient")
-            self.draw_labelled_prop(
-                box.row(), settings, ["env_map_fb_alpha"], "Use FB Alpha")
-
-    #######################################################
-    def draw_bump_map_box(self, context, box):
-
-        settings = context.material.dff
-        box.row().prop(settings, "export_bump_map")
-        
-        if settings.export_bump_map:
-            box.row().prop(settings, "bump_map_tex", text="Diffuse Texture")
-
-    #######################################################
-    def draw_uv_anim_box(self, context, box):
-
-        settings = context.material.dff
-
-        box.row().prop(settings, "export_animation")
-        if settings.export_animation:
-            box.row().prop(settings, "animation_name", text="Name")
-            
-            
-    #######################################################
-    def draw_refl_box(self, context, box):
-
-        settings = context.material.dff
-        box.row().prop(settings, "export_reflection")
-
-        if settings.export_reflection:
-            self.draw_labelled_prop(
-                box.row(), settings, ["reflection_scale_x", "reflection_scale_y"],
-                "Scale"
-            )
-            self.draw_labelled_prop(
-                box.row(), settings, ["reflection_offset_x", "reflection_offset_y"],
-                "Offset"
-            )
-            self.draw_labelled_prop(
-                box.row(), settings, ["reflection_intensity"], "Intensity"
-            )
-
-    #######################################################
-    def draw_specl_box(self, context, box):
-
-        settings = context.material.dff
-        box.row().prop(settings, "export_specular")
-
-        if settings.export_specular:
-            self.draw_labelled_prop(
-                box.row(), settings, ["specular_level"], "Level"
-            )
-            box.row().prop(settings, "specular_texture", text="Texture")
-        
-    #######################################################
-    def draw_mesh_menu(self, context):
-
-        layout = self.layout
-        settings = context.material.dff
-
-        layout.prop(settings, "ambient")
-
-        # This is for conveniently setting the base colour from the settings
-        # without removing the texture node
-        
-        try:
-
-            if bpy.app.version >= (2, 80, 0):
-                prop = context.material.node_tree.nodes["Principled BSDF"].inputs[0]
-                prop_val = "default_value"
-            else:
-                prop = context.material
-                prop_val = "diffuse_color"
-                
-            row = layout.row()
-            row.prop(
-                prop,
-                prop_val,
-                text="Color")
-            
-            row.prop(settings,
-                     "preset_mat_cols",
-                     text="",
-                     icon="MATERIAL",
-                     icon_only=True
-            )
-            
-        except Exception as e:
-            print(e)
-                
-        
-        self.draw_env_map_box  (context, layout.box())
-        self.draw_bump_map_box (context, layout.box())
-        self.draw_refl_box     (context, layout.box())
-        self.draw_specl_box    (context, layout.box())
-        self.draw_uv_anim_box  (context, layout.box())
-
-    #######################################################
-    # Callback function from preset_mat_cols enum
-    def set_preset_color(self, context):
-        try:
-            color = eval(context.material.dff.preset_mat_cols)
-            color = [i / 255 for i in color]
-                
-            if bpy.app.version >= (2, 80, 0):                
-                node = context.material.node_tree.nodes["Principled BSDF"]
-                node.inputs[0].default_value = color
-
-            # Viewport color in Blender 2.8 and Material color in 2.79.
-            context.material.diffuse_color = color[:-1]
-
-        except Exception as e:
-            print(e)
-        
-    #######################################################
-    def draw(self, context):
-
-        if not context.material or not context.material.dff:
-            return
-        
-        if context.object.dff.type == 'COL':
-            self.draw_col_menu(context)
-            return
-
-        self.draw_mesh_menu(context)
-
-#######################################################@
-class DFF_MT_ExportChoice(bpy.types.Menu):
-    bl_label = "DragonFF"
-
-    def draw(self, context):
-        self.layout.operator(EXPORT_OT_dff.bl_idname, text="DragonFF DFF (.dff)")
-        self.layout.operator(EXPORT_OT_col.bl_idname, text="DragonFF Collision (.col)")
-            
-        
-#######################################################
-def import_dff_func(self, context):
-    self.layout.operator(IMPORT_OT_dff.bl_idname, text="DragonFF DFF (.dff)")
-
-#######################################################
-def export_dff_func(self, context):
-    self.layout.menu("DFF_MT_ExportChoice", text="DragonFF")
-
-#######################################################
-class OBJECT_PT_dffObjects(bpy.types.Panel):
-
-    bl_idname      = "OBJECT_PT_dffObjects"
-    bl_label       = "DragonFF - Export Object"
-    bl_space_type  = "PROPERTIES"
-    bl_region_type = "WINDOW"
-    bl_context     = "object"
-
-    #######################################################
-    def draw_labelled_prop(self, row, settings, props, label, text=""):
-        
-        row.label(text=label)
-        for prop in props:
-            row.prop(settings, prop, text=text)
-
-
-    #######################################################
-    def validate_pipeline(self, pipeline):
-
-        try:
-            int(pipeline, 0)
-        except ValueError:
-            return False
-
-        return True
-            
-    #######################################################
-    def draw_mesh_menu(self, context):
-
-        layout = self.layout
-        settings = context.object.dff
-
-        box = layout.box()
-        box.prop(settings, "pipeline", text="Pipeline")
-        if settings.pipeline == 'CUSTOM':
-            col = box.column()
-            
-            col.alert = not self.validate_pipeline(settings.custom_pipeline)
-            icon = "ERROR" if col.alert else "NONE"
-
-            col.prop(settings, "custom_pipeline", icon=icon, text="Custom Pipeline")
-        box.prop(settings, "export_normals", text="Export Normals")
-        box.prop(settings, "export_binsplit", text="Export Bin Mesh PLG")
-
-            
-        properties = [         
-            ["day_cols", "Day Vertex Colours"],
-            ["night_cols", "Night Vertex Colours"],
-        ]
-
-        box = layout.box()
-        box.label(text="Export Vertex Colours")
-        
-        for property in properties:
-            box.prop(settings, property[0], text=property[1])
-
-        box = layout.box()
-        box.label(text="Export UV Maps")
-
-        box.prop(settings, "uv_map1", text="UV Map 1")
-
-        # Second UV Map can only be disabled if the first UV map is enabled.
-        if settings.uv_map1:
-            box.prop(settings, "uv_map2", text="UV Map 2")
-
-    #######################################################
-    def draw_col_menu(self, context):
-        layout = self.layout
-        settings = context.object.dff
-
-        box = layout.box()
-        box.label(text="Material Surface")
-        
-        box.prop(settings, "col_material", text="Material")
-        box.prop(settings, "col_flags", text="Flags")
-        box.prop(settings, "col_brightness", text="Brightness")
-        box.prop(settings, "col_light", text="Light")
-
-        pass
-            
-    #######################################################
-    def draw_obj_menu(self, context):
-
-        layout = self.layout
-        settings = context.object.dff
-
-        layout.prop(settings, "type", text="Type")
-
-        if settings.type == 'OBJ':
-            if context.object.type == 'MESH':
-                self.draw_mesh_menu(context)
-
-        elif settings.type == 'COL':
-            if context.object.type == 'EMPTY':
-                self.draw_col_menu(context)
-    
-    #######################################################
-    def draw(self, context):
-
-        if not context.object.dff:
-            return
-        
-        self.draw_obj_menu(context)
-    
-# Custom properties
-#######################################################
-class DFFMaterialProps(bpy.types.PropertyGroup):
-
-    ambient           : bpy.props.FloatProperty  (name="Ambient Shading", default=1)
-    
-    # Environment Map
-    export_env_map    : bpy.props.BoolProperty   (name="Environment Map")
-    env_map_tex       : bpy.props.StringProperty ()
-    env_map_coef      : bpy.props.FloatProperty  ()
-    env_map_fb_alpha  : bpy.props.BoolProperty   ()
-
-    # Bump Map
-    export_bump_map   : bpy.props.BoolProperty   (name="Bump Map")
-    bump_map_tex      : bpy.props.StringProperty ()
-
-    # Reflection
-    export_reflection    : bpy.props.BoolProperty  (name="Reflection Material")
-    reflection_scale_x   : bpy.props.FloatProperty ()
-    reflection_scale_y   : bpy.props.FloatProperty ()
-    reflection_offset_x  : bpy.props.FloatProperty ()
-    reflection_offset_y  : bpy.props.FloatProperty ()
-    reflection_intensity : bpy.props.FloatProperty ()
-    
-    # Specularity
-    export_specular  : bpy.props.BoolProperty(name="Specular Material")
-    specular_level   : bpy.props.FloatProperty  ()
-    specular_texture : bpy.props.StringProperty ()
-
-    # Collision Data
-    col_flags       : bpy.props.IntProperty()
-    col_brightness  : bpy.props.IntProperty()
-    col_light       : bpy.props.IntProperty()
-    col_mat_index   : bpy.props.IntProperty()
-
-    # UV Animation
-    export_animation : bpy.props.BoolProperty   (name="UV Animation")
-    animation_name   : bpy.props.StringProperty ()
-
-    # Pre-set Material Colours
-    preset_mat_cols : bpy.props.EnumProperty(
-        items =
-        (
-            ("[255, 60, 0, 255]", "Right Tail Light", ""),
-            ("[185, 255, 0, 255]", "Left Tail Light", ""),
-            ("[0, 255, 200, 255]", "Right Headlight", ""),
-            ("[255, 175, 0, 255]", "Left Headlight", ""),
-            ("[0, 255, 255, 255]", "4 Colors Paintjob", ""),
-            ("[255, 0, 175, 255]", "Secondary Color", ""),
-            ("[60, 255, 0, 255]", "Primary Color", ""),
-            ("[184, 255, 0, 255]", "ImVehFT - Breaklight L", ""),
-            ("[255, 59, 0, 255]", "ImVehFT - Breaklight R", ""),
-            ("[255, 173, 0, 255]", "ImVehFT - Revlight L", ""),
-            ("[0, 255, 198, 255]", "ImVehFT - Revlight R", ""),
-            ("[255, 174, 0, 255]", "ImVehFT - Foglight L", ""),
-            ("[0, 255, 199, 255]", "ImVehFT - Foglight R", ""),
-            ("[183, 255, 0, 255]", "ImVehFT - Indicator LF", ""),
-            ("[255, 58, 0, 255]", "ImVehFT - Indicator RF", ""),
-            ("[182, 255, 0, 255]", "ImVehFT - Indicator LM", ""),
-            ("[255, 57, 0, 255]", "ImVehFT - Indicator RM", ""),
-            ("[181, 255, 0, 255]", "ImVehFT - Indicator LR", ""),
-            ("[255, 56, 0, 255]", "ImVehFT - Indicator RR", ""),
-            ("[0, 16, 255, 255]", "ImVehFT - Light Night", ""),
-            ("[0, 17, 255, 255]", "ImVehFT - Light All-day", ""),
-            ("[0, 18, 255, 255]", "ImVehFT - Default Day", "")
-        ),
-        update = MATERIAL_PT_dffMaterials.set_preset_color
-    )
-    
-    def register():
-        bpy.types.Material.dff = bpy.props.PointerProperty(type=DFFMaterialProps)
-        
-#######################################################
-class DFFObjectProps(bpy.types.PropertyGroup):
-
-    # Atomic Properties
-    type : bpy.props.EnumProperty(
-        items = (
-            ('OBJ', 'Object', 'Object will be exported as a mesh or a dummy'),
-            ('COL', 'Collision Object', 'Object is a collision object'),
-            ('SHA', 'Shadow Object', 'Object is a shadow object'),
-            ('NON', "Don't export", 'Object will NOT be exported.')
-        )
-    )
-
-    # Mesh properties
-    pipeline : bpy.props.EnumProperty(
-        items = (
-            ('NONE', 'None', 'Export without setting a pipeline'),
-            ('0x53F20098', 'Buildings', 'Refl. Building Pipleine (0x53F20098)'),
-            (
-                '0x53F2009A',
-                'Night Vertex Colors',
-                'Night Vertex Colors (0x53F2009C)'
-            ),
-            ('CUSTOM', 'Custom Pipeline', 'Set a different pipeline')
-        ),
-        name="Pipeline",
-        description="Select the Engine rendering pipeline"
-    )
-    custom_pipeline : bpy.props.StringProperty(name="Custom Pipeline")
-    
-    export_normals : bpy.props.BoolProperty(
-        default=True,
-        description="Whether Normals will be exported. (Disable for Map objects)"
-    )
-    
-    uv_map1 : bpy.props.BoolProperty(
-        default=True,
-        description="First UV Map will be exported")
-    
-    uv_map2 : bpy.props.BoolProperty(
-        default=True,
-        description="Second UV Map will be exported"
-    )
-    
-    day_cols : bpy.props.BoolProperty(
-        default=True,
-        description="Whether Day Vertex Prelighting Colours will be exported"
-    )
-    
-    night_cols : bpy.props.BoolProperty(
-        default=True,
-        description="Extra prelighting colours. (Tip: Disable export normals)"
-    )
-    
-    export_binsplit : bpy.props.BoolProperty(
-        default=True,
-        description="Enabling will increase file size, but will increase\
-compatibiility with DFF Viewers"
-    )
-
-    col_material : bpy.props.IntProperty(
-        default = 12,
-        description = "Material used for the Sphere/Cone"
-    )
-
-    col_flags : bpy.props.IntProperty(
-        default = 0,
-        description = "Flags for the Sphere/Cone"
-    )
-
-    col_brightness : bpy.props.IntProperty(
-        default = 0,
-        description = "Brightness used for the Sphere/Cone"
-    )
-    
-    col_light : bpy.props.IntProperty(
-        default = 0,
-        description = "Light used for the Sphere/Cone"
-    )
-    
-    #######################################################    
-    def register():
-        bpy.types.Object.dff = bpy.props.PointerProperty(type=DFFObjectProps)
-
-#######################################################
-class DFFSceneProps(bpy.types.PropertyGroup):
-
-    #######################################################    
-    def update_map_sections(self, context):
-        return map_data.data[self.game_version_dropdown]['IPL_paths']
-        
-    game_version_dropdown : bpy.props.EnumProperty(
-        name = 'Game',
-        items = (
-            (game_version.III, 'GTA III', 'GTA III map segments'),
-            (game_version.VC, 'GTA VC', 'GTA VC map segments'),
-            (game_version.SA, 'GTA SA', 'GTA SA map segments'),
-            (game_version.LCS, 'GTA LCS', 'GTA LCS map segments'),
-            (game_version.VCS, 'GTA VCS', 'GTA VCS map segments'),
-        )
-    )
-
-    map_sections : bpy.props.EnumProperty(
-        name = 'Map segment',
-        items = update_map_sections
-    )
-
-    skip_lod: bpy.props.BoolProperty(
-        name        = "Skip LOD Objects",
-        default     = False
-    )
-
-    game_root : bpy.props.StringProperty(
-        name = 'Game root',
-        default = 'C:\\Program Files (x86)\\Steam\\steamapps\\common\\',
-        description = "Folder with the game's executable",
-        subtype = 'DIR_PATH'
-    )
-
-    dff_folder : bpy.props.StringProperty(
-        name = 'Dff folder',
-        default = 'C:\\Users\\blaha\\Documents\\GitHub\\DragonFF\\tests\\dff',
-        description = "Define a folder where all of the dff models are stored.",
-        subtype = 'DIR_PATH'
-    )
-
-    # txd_folder = bpy.props.StringProperty \
-    #     (
-    #     name = 'Txd folder',
-    #     default = 'C:\\Users\\blaha\\Documents\\GitHub\\DragonFF\\tests\\txd',
-    #     description = "Define a folder where all of the txd models are stored.",
-    #     subtype = 'DIR_PATH'
-    #     )
-
-    #######################################################    
-    def register():
-        bpy.types.Scene.dff = bpy.props.PointerProperty(type=DFFSceneProps)
-
-#######################################################
-class MapImportPanel(bpy.types.Panel):
-    """Creates a Panel in the scene context of the properties editor"""
-    bl_label = "DragonFF - Map Import"
-    bl_idname = "SCENE_PT_layout"
-    bl_space_type = 'PROPERTIES'
-    bl_region_type = 'WINDOW'
-    bl_context = "scene"
-
-    #######################################################
-    def draw(self, context):
-        layout = self.layout
-        settings = context.scene.dff
-
-        flow = layout.grid_flow(row_major=True,
-                                columns=0,
-                                even_columns=True,
-                                even_rows=False,
-                                align=True)
-
-        col = flow.column()
-        col.prop(settings, "game_version_dropdown", text="Game")
-        col.prop(settings, "map_sections", text="Map segment")
-        col.separator()
-        col.prop(settings, "skip_lod", text="Skip LOD objects")
-
-        layout.separator()
-
-        layout.prop(settings, 'game_root')
-        layout.prop(settings, 'dff_folder')
-
-        row = layout.row()
-        row.operator("scene.dragonff_map_import")
