+++ gtaLib/map.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from . import map_data
+from ..data import map_data
 from collections import namedtuple
 
 Vector = namedtuple("Vector", "x y z")
