+++ ops/dff_exporter.py
@@ -20,10 +20,8 @@ import mathutils
 import os
 import os.path
 
-from . import dff
+from ..gtaLib import dff
 from .col_exporter import export_col
-from . import col
-from .importer_common import set_object_mode
 
 #######################################################
 def clear_extension(string):
@@ -187,6 +185,14 @@ class material_helper:
             props.reflection_intensity
         )
 
+    #######################################################
+    def get_user_data(self):
+
+        if 'dff_user_data' not in self.material:
+            return None
+        
+        return dff.UserData.from_mem(
+                self.material['dff_user_data'])
     
     #######################################################
     def get_uv_animation(self):
@@ -319,6 +325,9 @@ class dff_exporter:
             matrix.to_3x3().transposed()
         )
 
+        if "dff_user_data" in obj:
+            frame.user_data = dff.UserData.from_mem(obj["dff_user_data"])
+
         id_array = self.bones if is_bone else self.frames
         
         if set_parent and obj.parent is not None:
@@ -357,6 +366,7 @@ class dff_exporter:
             material.add_plugin('env_map', helper.get_environment_map())
             material.add_plugin('spec', helper.get_specular_material())
             material.add_plugin('refl', helper.get_reflection_material())
+            material.add_plugin('udata', helper.get_user_data())
 
             anim = helper.get_uv_animation()
             if anim:
@@ -641,6 +651,9 @@ class dff_exporter:
             geometry.extensions['skin'] = skin
         if extra_vert:
             geometry.extensions['extra_vert_color'] = extra_vert
+        if "dff_user_data" in obj.data:
+            geometry.extensions['user_data'] = dff.UserData.from_mem(
+                obj.data['dff_user_data'])
 
         try:
             if obj.dff.pipeline != 'NONE':
